+++ fissure/Dashboard/UI_Components/UI_Types.py
@@ -19,5 +19,6 @@ class UI_Types:
     JointPlot, _ = uic.loadUiType(os.path.join(fissure.utils.UI_DIR, "joint_plot.ui"))
     Trim, _ = uic.loadUiType(os.path.join(fissure.utils.UI_DIR, "trim.ui"))
     Triggers, _ = uic.loadUiType(os.path.join(fissure.utils.UI_DIR, "triggers.ui"))
+    Demod, _ = uic.loadUiType(os.path.join(fissure.utils.UI_DIR, "demod.ui"))
     Features, _ = uic.loadUiType(os.path.join(fissure.utils.UI_DIR, "features.ui"))